// frontend/magic-web/packages/magic-admin/src/provider/AdminProvider/types.ts
export const enum AppEnv {
	/** 生产环境 */
	Production = "saas-prod",
	/** 预发布环境 */
	Pre = "saas-pre",
	/** 测试环境 */
	Test = "saas-test",
	/** 预发布国际环境 */
	InternationalPre = "international-pre",
	/** 国际版 生产环境 */
	InternationalProduction = "international-prod",
}

/** 全局国际冠号选项 */
export interface AreaCodeOption {
	/** 当前语言标识 */
	locale: string
	/** 当前冠号识别号 */
	code: string
	/** 当前语言名称 */
	name: string
	/** 用于当前语言的枚举各语言的表达 */
	translations: Record<string, string>
}

/** 用户信息 */
export interface UserInfo {
	// teamshare 用户id
	id?: string
	// Magic 用户id
	magic_id: string
	user_id: string
	status: string | number
	nickname: string
	real_name: string
	avatar: string
	organization_code: string
	phone?: string
	email?: string
	country_code?: string
}

/** Teamshare 用户信息 */
export interface TeamshareUserInfo {
	id: string
	name: string
	avatar: string
	departments: {
		name: string
		id: string
	}[][]
}

/**
 * API 客户端配置
 */
export interface ApiClients {
	/** Magic 服务请求实例 */
	magicClient: HttpClient
}
// ...
export type EnhancedNavigateFunction = Partial<RouteParams> &
	EnhancedNavigateOptions & {
		/** history.go */
		delta?: number
	}
// ...
export interface Organization {
	organizationCode: string
	teamshareOrganizationCode: string
	organizationInfo: MagicOrganization | null
	teamshareOrganizationInfo: TeamshareOrganization | null
}

/**
 * 管理后台桥接配置
 * 用于从 magic 项目注入依赖
 */
export interface AdminBridgeConfig {
	/** API 客户端实例（与 magic 共享） */
	apiClients: ApiClients
	/** 环境变量 */
	env: {
		MAGIC_APP_ENV: AppEnv
		MAGIC_BASE_URL: string
	}
	/** 用户信息 （与 magic 共享 */
	user: {
		/** 用户Token（与 magic 共享） */
		token: string | null
		/** 用户信息 （与 magic 共享 */
		userInfo: UserInfo | null
	}
	/** 当前账户所处的组织信息 （与 magic 共享 */
	organization: Organization
	/** 路由导航函数（与 magic 共享） */
	navigate: (props: EnhancedNavigateFunction) => void
	/** 路由导航组件 */
	Navigate?: React.ComponentType<NavigateProps>
	/** 集群编码 */
	clusterCode?: string
	/** 全局国际冠号选项 */
	areaCodes?: AreaCodeOption[] | null
	/** 是否是私有部署 */
	isPrivateDeployment?: boolean
	/** 安全区域间距 */
	safeAreaInset?: {
		bottom?: number | string
		top?: number | string
	}
	/** 其他自定义配置 */
	[key: string]: any
}

export interface AdminProviderProps extends AdminBridgeConfig {
	/** 主题 */
	theme?: "light" | "dark"
	/** 语言 */
	language?: LanguageType
}

export interface AdminProviderContextType extends AdminProviderProps {
	getLocale: <T extends keyof LocaleType>(namespace: T) => LocaleType[T]
}